scout/requirements.txt
psutil>=5.9
pywin32>=307
websockets==10.4          # 🔶 pin to 10.x where extra_headers existed
python-dotenv>=1.0
win10toast>=0.9
